# docs/0001-api-openapi.yaml
openapi: 3.0.3
info:
  title: Peluquería — Gestión de Trabajadores
  version: 1.0.0
  description: API para alta de trabajadores, fichajes de jornada y solicitudes de vacaciones.
servers:
  - url: /api
paths:
...
  /workers:
    get:
      summary: Listar trabajadores
      tags: [Workers]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de trabajadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Worker'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Alta de un trabajador
      tags: [Workers]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerCreateRequest'
      responses:
        '201':
          description: Trabajador creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workers/status:
    get:
      summary: Estado actual de todos los trabajadores
      tags: [Workers]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de trabajadores con su estado actual
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkerStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workers/login:
    post:
      summary: Iniciar sesión por PIN
      tags: [Workers]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Estado del trabajador autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workers/{id}/status:
    get:
      summary: Estado actual de un trabajador por id
      tags: [Workers]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Estado del trabajador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workers/{id}/change-pin:
    post:
      summary: Cambiar el PIN de un trabajador
      tags: [Workers]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePinRequest'
      responses:
        '204':
          description: PIN actualizado
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Sin permisos (rol insuficiente)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Role:
      type: string
      enum: [admin, worker]
...
    WorkerCreateRequest:
      type: object
      required: [fullName, pin, role]
      properties:
        fullName:
          type: string
        pin:
          type: string
          pattern: '^\d{4}$'
        role:
          $ref: '#/components/schemas/Role'
    LoginRequest:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
          pattern: '^\d{4}$'
    ChangePinRequest:
      type: object
      required: [currentPin, newPin]
      properties:
        currentPin:
          type: string
          pattern: '^\d{4}$'
        newPin:
          type: string
          pattern: '^\d{4}$'
    Worker:
      type: object
      required: [id, fullName, role]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        active:
          type: boolean
    WorkerStatus:
      type: object
      required: [id, fullName, role, currentStatus]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        currentStatus:
          type: string
          enum: [idle, in-progress, paused, finished]
        startedAt:
          type: string
          format: date-time
        currentPausedAt:
          type: string
          format: date-time
        workedTime:
          type: string
          description: Duración en formato hh:mm:ss
        pausedTime:
          type: string
          description: Duración en formato hh:mm:ss
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TimelineSegment'
...
    TimelineSegment:
      type: object
      required: [type, from, to]
      properties:
        type:
          type: string
          enum: [working, paused]
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string